refactor(scroll-stories): move gesture indicator to separate component
