+++ src/scripts/components/stories/story/scroll-story/story-eei/splashscreen/splashscreen-eei.module.css
@@ -38,32 +38,4 @@
       fill: none;
     }
   }
-
-  .gestureIndicator {
-    position: absolute;
-    bottom: 100px;
-    left: 50%;
-    transform: translateX(-50%);
-    color: var(--main-colours-esa-teal-55);
-    display: grid;
-    place-content: center;
-    margin: 0 auto;
-
-    &:after {
-      color: var(--main-colours-esa-teal-55);
-      text-align: center;
-      font-family: Arial;
-      font-size: 0.75rem;
-      font-weight: 700;
-      line-height: 18px; /* 150% */
-      letter-spacing: 1.5px;
-      text-transform: uppercase;
-      bottom: -2rem;
-      left: 50%;
-      white-space: nowrap;
-      transform: translateX(-50%);
-      position: absolute;
-      content: attr(data-content);
-    }
-  }
 }
