+++ src/scripts/components/stories/story/scroll-story/story-x-fires/splashscreen/splashscreen-x-fires.module.css
@@ -35,30 +35,4 @@
       }
     }
   }
-
-  .gestureIndicator {
-    position: absolute;
-    bottom: 100px;
-    left: 50%;
-    transform: translateX(-50%);
-    display: grid;
-    place-content: center;
-    margin: 0 auto;
-
-    &:after {
-      text-align: center;
-      font-family: Arial;
-      font-size: 0.75rem;
-      font-weight: 700;
-      line-height: 18px; /* 150% */
-      letter-spacing: 1.5px;
-      text-transform: uppercase;
-      bottom: -2rem;
-      left: 50%;
-      white-space: nowrap;
-      transform: translateX(-50%);
-      position: absolute;
-      content: attr(data-content);
-    }
-  }
 }
